jets: rename `maps` to `seam`
