+++ rust/ares/src/jets.rs
@@ -8,10 +8,10 @@ pub mod hash;
 pub mod list;
 pub mod lock;
 pub mod lute;
-pub mod maps;
 pub mod math;
 pub mod nock;
 pub mod parse;
+pub mod seam;
 pub mod serial;
 pub mod sort;
 pub mod tree;
@@ -28,10 +28,10 @@ use crate::jets::lock::aes::*;
 use crate::jets::lock::ed::*;
 use crate::jets::lock::sha::*;
 use crate::jets::lute::*;
-use crate::jets::maps::*;
 use crate::jets::math::*;
 use crate::jets::nock::*;
 use crate::jets::parse::*;
+use crate::jets::seam::*;
 use crate::jets::serial::*;
 use crate::jets::sort::*;
